+++ test/cases/js-import-image-svg-inline/src/main.js
@@ -2,7 +2,9 @@ import svgFilename from './token.svg'; // import as output filename
 import svgBase64DataUrl from './token.svg?inline=base64'; // import as base64-encoded data URL
 import svgUtf8DataUrl from './token.svg?inline'; // import as UTF-8 data URL
 import svgUtf8DataUrl2 from './token.svg?inline=utf8'; // import as UTF-8 data URL
-//import svgSource from './token.svg?raw'; // import as source content
+
+import pngFilename from '@images/icons/imac.png';
+import pngDataUrl from '@images/icons/imac.png?inline';
 
 // filename
 let img1 = document.createElement('img');
@@ -23,3 +25,7 @@ console.log('>> svgFilename: ', svgFilename);
 console.log('>> svgBase64DataUrl: ', svgBase64DataUrl);
 console.log('>> svgUtf8DataUrl: ', svgUtf8DataUrl);
 console.log('>> svgUtf8DataUrl2: ', svgUtf8DataUrl2);
+
+// test bin images
+console.log('>> pngFilename: ', pngFilename);
+console.log('>> pngDataUrl: ', pngDataUrl);